moved the packet class code out of drx.py
